Header image can now be set from main pages article image
